refactor(api): auth → src/api modules (Effort 1)
Move the auth data-access slice from src/hooks/queries/auth to a
feature-sliced src/api/auth module: shared types.ts (Profile Row + key
factories) plus one use-prefixed file per endpoint, each with a per-file
queryOptions factory. No behavior change; all consumer imports repointed.
