+++ src/api/auth/useUpdateProfile.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useToast } from "@/hooks/use-toast";
 import { supabase } from "@/integrations/supabase/client";
-import { profileKeys } from "./useProfile";
+import { profileKeys } from "./types";
 
 // Mutation function
 async function updateProfile(variables: {
