+++ src/components/layout/AppHeader/UserMenu.tsx
@@ -12,7 +12,7 @@ import { Button } from "@/components/ui/button";
 import { UserAvatar } from "./UserAvatar";
 import { Database } from "@/integrations/supabase/types";
 import { Link } from "@tanstack/react-router";
-import { useUserPermissionsQuery } from "@/hooks/queries/auth/useUserPermissions";
+import { useUserPermissionsQuery } from "@/api/auth/useUserPermissions";
 
 type Profile = Database["public"]["Tables"]["profiles"]["Row"];
