+++ src/pages/admin/AddGenreDialog.tsx
@@ -12,7 +12,7 @@ import {
 } from "@/components/ui/dialog";
 import { Plus } from "lucide-react";
 
-import { useUserPermissionsQuery } from "@/hooks/queries/auth/useUserPermissions";
+import { useUserPermissionsQuery } from "@/api/auth/useUserPermissions";
 import { useAuth } from "@/contexts/AuthContext";
 
 interface AddGenreDialogProps {
