+++ src/pages/groups/Groups.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { useNavigate } from "@tanstack/react-router";
 import { useAuth } from "@/contexts/AuthContext";
 import { useUserGroupsQuery } from "@/hooks/queries/groups/useUserGroups";
-import { useUserPermissionsQuery } from "@/hooks/queries/auth/useUserPermissions";
+import { useUserPermissionsQuery } from "@/api/auth/useUserPermissions";
 import { useDeleteGroupMutation } from "@/hooks/queries/groups/useDeleteGroup";
 import { DeleteGroupDialog } from "./Groups/DeleteGroupDialog";
 import { CreateGroupDialog } from "./Groups/CreateGroupDialog";
